+++ lib/twilio-ruby/rest/verify/v2/service/verification.rb
@@ -49,6 +49,7 @@ module Twilio
                     # @param [String] template_sid The message [template](https://www.twilio.com/docs/verify/api/templates). If provided, will override the default template for the Service. SMS and Voice channels only.
                     # @param [String] template_custom_substitutions A stringified JSON object in which the keys are the template's special variables and the values are the variables substitutions.
                     # @param [String] device_ip Strongly encouraged if using the auto channel. The IP address of the client's device. If provided, it has to be a valid IPv4 or IPv6 address.
+                    # @param [Boolean] enable_sna_client_token An optional Boolean value to indicate the requirement of sna client token in the SNA URL invocation response for added security. This token must match in the Verification Check request to confirm phone number verification.
                     # @param [RiskCheck] risk_check 
                     # @param [String] tags A string containing a JSON map of key value pairs of tags to be recorded as metadata for the message. The object may contain up to 10 tags. Keys and values can each be up to 128 characters in length.
                     # @return [VerificationInstance] Created VerificationInstance
@@ -68,6 +69,7 @@ module Twilio
                         template_sid: :unset, 
                         template_custom_substitutions: :unset, 
                         device_ip: :unset, 
+                        enable_sna_client_token: :unset, 
                         risk_check: :unset, 
                         tags: :unset
                     )
@@ -88,6 +90,7 @@ module Twilio
                             'TemplateSid' => template_sid,
                             'TemplateCustomSubstitutions' => template_custom_substitutions,
                             'DeviceIp' => device_ip,
+                            'EnableSnaClientToken' => enable_sna_client_token,
                             'RiskCheck' => risk_check,
                             'Tags' => tags,
                         })